Started modifying sight data
